Move RH, Wind and Tas converters to new file in indices and create indicators Tests still needed.
